+++ application/tests/models/Sites_Model_test.php
@@ -470,6 +470,20 @@ class Sites_Model_test extends TestCase {
 		$this->_testSiteFailure('WhiteCloudPavillion', 'Bad Status Code (404)');
 	}
 
+	public function test_WorldThree() {
+		$testSeries = [
+			'asayake_wa_koganeiro'         => 'Asayake wa Koganeiro',
+			'black_bullet'                 => 'Black Bullet',
+			'dont_cry_girl'                => 'Don\'t Cry, Girl'
+		];
+		$randSeries = array_rand($testSeries);
+
+		$this->_testSiteSuccess('WorldThree', $randSeries, $testSeries[$randSeries]);
+	}
+	public function test_WorldThree_fail() {
+		$this->_testSiteFailure('WorldThree', 'Bad Status Code (404)');
+	}
+
 	private function _testSiteSuccess(string $siteName, string $title_url, string $expectedTitle) {
 		$result = $this->Sites_Model->{$siteName}->getTitleData($title_url);