+++ application/views/Help.php
@@ -54,6 +54,7 @@
 		<li>FallenAngelsScans</li>
 		<li>Mangaichi Scanlation Division</li>
 		<li>LHtranslation</li>
-		<li>White Cloud Pavillion</li>
+		<li>White Cloud Pavilion</li>
+		<li>World Three</li>
 	</ul>
 </div>