+++ public/userscripts/manga-tracker.user.js
@@ -35,8 +35,9 @@
 // @include      /^http?:\/\/mangaichiscans\.mokkori\.fr\/fs\/read\/.*?\/[a-z]+\/[0-9]+\/[0-9]+(\/.*)?$/
 // @include      /^http:\/\/lhtranslation\.com\/read-(.*?)-chapter-[0-9\.]+\.html$/
 // @include      /^https?:\/\/archangelscans\.com\/free\/read\/.*?\/[a-z]+\/[0-9]+\/[0-9]+(\/.*)?$/
+// @include      /^http:\/\/www\.slide\.world-three\.org\/read\/.*?\/[a-z]+\/[0-9]+\/[0-9]+(\/.*)?$/
 // @updated      2017-08-17
-// @version      1.7.52
+// @version      1.7.53
 // @downloadURL  https://trackr.moe/userscripts/manga-tracker.user.js
 // @updateURL    https://trackr.moe/userscripts/manga-tracker.meta.js
 // @require      https://ajax.googleapis.com/ajax/libs/jquery/3.1.1/jquery.min.js
@@ -2167,6 +2168,17 @@ let sites = {
 		preInit : function(callback) {
 			this.foolSlideBaseURL = this.https+'://archangelscans.com/free';
 			this.setupFoolSlide(3);
+			callback();
+		}
+	}),
+
+	/**
+	 * World Three (FoolSlide)
+	 * @type {SiteObject}
+	 */
+	'www.slide.world-three.org' : extendSite({
+		preInit : function(callback) {
+			this.setupFoolSlide(2);
 			callback();
 		}
 	}),